Allow the reuse of an event loop by an OFClient.
This commit enables more than one client in an event loop.
It adds a pointer to an Eventloop object to the constructor
of BaseOFClient. The BaseOFClient constructor
checks if the Eventloop is valid and assigns it to its class
member evloop.
The commit also adds an option in OFServerSettings to pass
an Eventloop to the constructor of OFClient. It then passes
the Eventloop to its parent constructor.